Fix "cross-compiling" `ar`/`strip` when building in MSYS2
No idea why, but the mingw-w64-gcc package calls the `ar` executable
with target prefix `x86_64-w64-mingw32-gcc-ar`, i.e. with a spurious
`gcc-` thrown in. Whatevs. Just work around it and move on.

Likewise, there simply does not exist any `strip.exe` with the
`x86_64-w64-mingw32-` prefix, as the `strip` executable is installed via
the `mingw-w64-binutils` package (which does not bother with hard
linking or copying the `.exe` files to their prefixed counterparts).

Of course, we only want to override this when building in MSYS2 (or,
more precisely, when using the mingw-w64 compiler). To that end,
guard the override behind a test for the specific `$(AR)` value.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>