copyfd: reinstate proper guard around munmap()
In 4c1392296 (Introduce FEATURE_COPYBUF_KB, 2007-12-02), a feature was
introduced where a large stack was allocated via mmap(), and
consequently released via munmap(). Since this is overkill for small
stacks, the mmap()/munmap() code was guarded inside an #if
CONFIG_FEATURE_COPYBUF_KB > 4 ... #endif guard.

In pure Win32 API, there is no support to abuse mmap() to allocate a
stack that way, therefore the code is simply disabled in busybox-w32.

But 8d75d794e (libbb: use sendfile() to copy data between file
descriptors, 2014-11-27) removed that guard around the munmap() call.
That was most likely a mistake, as the corresponding mmap() call is
*still* inside an equivalent #if ... #endif guard.

Let's revert the mistaken change.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>