inet_common: handle FEATURE_ETC_NETWORKS as a compile-time flag
This flag is used in #if ... #endif guards elsewhere and there is no
good reason to pretend that it is a runtime flag in just one case.

This probably was an oversight in a61cb92f2 (make /etc/network parsing
configurable. -200 bytes when off., 2007-06-19).

Besides, it fixes a compile problem on windows where the getnetbyaddr()
function is not available.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>