ash: use JOBS as the compile time option it actually is
Using the constant in a regular `if (...)` pretends that this is
anything but a compile time option. Let's use `#if JOBS` instead,
making it *much* clearer what `JOBS` actually is.

Incidentally, this change fixes the build when JOBS is 0 because
of lack of support in the headers (and hence WIFSTOPPED and WSTOPSIG
missing).

This partially reverts 4700fb5be (ash: make dowait() a bit more
readable. Logic is unchanged, 2015-10-09).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>