sh: support BASH_XTRACEFD when built Bash compatibly
There is a very useful feature in Bash where you can redirect the trace
enabled by `set -x` to a file descriptor *different* than 2. This comes
in particularly handy when validating the error output of commands, say,
in Git's test suite, while tracing at the same time.

It is such a useful feature, and very easily implemented.

Note: this patch has to adjust the way REDIR_SAVEFD2 is handled, as
it hard-coded the fd 2 before.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>