HACK: ash: expand PS4 like PS1
This is a total HACK. It *should* work alright, most of the time, but it
does suffer from a curious way how read_input_line() includes the code
to initialize the global state, and while we need the global state (e.g.
to expand the working directory via \w because we need to know the
current user's home directory), we do not want to run read_input_line(),
obviously.

Therefore, this code would need to be refactored substantially, to
carefully determine what needs to be initialized once, and once only
(read_input_line() reinitializes *every single time*), then factor out a
separate function to initialize that state, figure out where this can be
safely deinitialized, and then also use that code in the xflag code
block in ash.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>