win32: implement symlink(2)
Provide an implementation of symlink(2).

Calls to symlink(2) will fail in default Windows installations unless
running with elevated privileges.  Failure to create a symlink when
extracting files from an archive is therefore treated as a non-fatal
error.

There are two ways to permit the creation of symlinks:

- Edit security policy to give users the 'Create symbolic links'
  privilege.  Unfortunately this doesn't work for users who are an
  Administrator.

- Enable developer mode, which is available in later versions of
  Windows 10.

The ability to create symlinks is not available in Windows XP
or ReactOS.