HACK: work around mingw-w64-gcc v10.2.0 issue
When using `gcc -r -o <combined>.o <object>.o` where `<object>.o` was
produced by `windres, mingw-w64-gcc v10.2.0 corrupts the contents so
that the final link fails with a couple of these messages:

	/mingw64/x86_64-w64-mingw32/bin/ld: error: 0-bit reloc in dll

The symptom was reported to the BusyBox-w32 project already, in
https://github.com/rmyorston/busybox-w32/issues/200

The worst part is that this `gcc -r -o` command is totally pointless
because we only have one `resources.o` in win32/resources/, there is no
other object file to be combined into a single one.

As a real ugly work-around, filter out the `<combined>.o` argument from
the link command, and add `resources.o` directly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>