win32: workaround for '0-bit reloc' on MSYS2
There seems to be a bug in the Windows/MSYS2/mingw-w64 toolchain
which results in the generated win32/resources/built-in.o lacking
relocation data.

Forcing an actual C program to be linked along with the resources
appears to be an effective workaround.

When the toolchain bug is resolved this commit can be reverted.

See GitHub issue #200.