which: do report absolute path for busybox.exe
When running `which.exe busybox`, we do want to print the absolute path
of `busybox.exe`, not the bare basename `busybox`.

This is actually required for `./runtest busybox` to pass, as it will
first link `busybox.exe` to `runtest-tempdir-links/which.exe` and then
run `$lncmd "$(which busybox)" unknown$suffix`, expecting that `which
busybox` to report the actual path to `busybox.exe`.

This commit reverts 407e31172 (which: explicit test for busybox when
applets are preferred, 2016-05-19) whose intention was left somewhat
unclear.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>