WIP ci: also run the test suite
TODO: fix `sed` tests, too.

Now that we added fixes or workarounds for all failures when running
with MSYS2, let's run it in CI builds, to keep it in this shape.

Note that we require a couple work-arounds, e.g. setting `tsdir` to a
Windows path (MSYS2 pretends some sort of Unix path that BusyBox won't
understand, e.g. `/c/Windows` instead of `C:\Windows`).

As to the question why we would use MSYS2's Bash to run the test suite
instead of BusyBox itself? That would not verify BusyBox, not really.
For example, if there was a bug in BusyBox' `find`, comparing the output
of `find [...]` with the output of `busybox find [...]` would not catch
that bug at all: the outputs would be identical because we would have
used BusyBox to generate both.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>