tests: be verbose only for tests that failed
With this change, in new-style tests we redirect stdout/stderr to a file
and reveal the contents only when a test case failed.

This imitates the way things are done in old-style tests, and allows for
really pleasant output with `-v` when all tests are passing (and it
makes it easier to find the relevant parts of the entire testsuite's
output when a test case failed).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>